pyproject.toml
[build-system]
requires = ['setuptools', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'django-queryset-feeler'
description = 'tools to help understand how django querysets are executed'
version = '0.0.5'
authors = [
    {name='Lukas Schillinger', email='lukas@schillingertools.com'}, 
]
readme = 'README.md'
requires-python = '>=3.7'
keywords = ['django', 'query', 'queryset']
license = {file= 'LICENSE'}
classifiers = [
    'Programming Language :: Python :: 3',
    'Operating System :: OS Independent',
    'Framework :: Django',
]
dependencies = [
    'sqlparse >= 0.4.2', 
    'django >= 3.2', 
    'sql-metadata >= 2.6.0', 
    'Pygments >= 2.12.0', 
    'ipython >= 8.4.0', 
]

[project.urls]
homepage = 'https://github.com/lukas-schillinger/django-queryset-feeler'

[project.optional-dependencies]
test = ['djangorestframework >= 3.13.1']